Adjust USB interrupts, reset, re-enable USB log
Logging over USB has been pretty flaky, device enumeration often not
working. There have been hints that the USB interrupt isn't serviced
quickly enough, but also about the interrupt not happening at all.

Trying to fix this by properly setting up interrupt priorities so that
USB can preempt everything else and get its shit done quickly. Trying
to reset the USB peripheral more thoroughly, and trying to not block
in _write() when the host isn't polling us (which happens when the
port is closed).

This seems to work, but the nature of USB is that you never know if it
still works tomorrow.