Include Kiss FFT
This is the Kiss FFT library (sourceforge.net/projects/kissfft). After
testing libfftw3 and getting annoyed at how hard it is to use and
build and how enormous it is (600KiB!!!), I found Kiss FFT. It is
slightly faster on Cortex-M3, and very easy to use and build. BSD
license.