USB audio: send the right number of samples each frame
We run in asynchronous mode and send the number of samples that are
actually available per USB frame to the host. How many samples have
trickled in since the last USB frame is deduced by looking at where
the DMA job writes to the input buffer when it's time to fill the USB
buffer.

I think there may be artefacts when starting to send audio; the code
for starting to sync up DMA and USB buffers is naive (well, it doesn't
really exist).