Disable automatic high-pass in ADC, doing it in software
This fixes some noise issues with the WM8731 codec, for some weird
reasons. The ADC really doesn't like it when it's quiet, when it's
sitting just around digital 0, and it seems to start to invent
noise. Disabling removing of the DC bias (an offset which seems to be
internal to the codec anyway) makes the signal much cleaner when the
input is absolutely quiet. The DC bias is now removed in software
instead.

When testing this I have noticed that it makes a difference what the
DAC does too, even if it is muted or at a very low volume level. It
doesn't like absolute silence either, in which case the ADC produces
some noise tones at a few tens of Hz.